+++ projects/ngx-amap/src/ngx-amap.module.ts
@@ -29,6 +29,7 @@ import { AmapEllipseDirective } from './directives/amap-ellipse/amap-ellipse.dir
 import { AmapRectangleDirective } from './directives/amap-rectangle/amap-rectangle.directive';
 import { AmapHeatmapDirective } from './directives/amap-heatmap/amap-heatmap.directive';
 import { AmapDrivingService, AmapDrivingWrapper } from './services/amap-driving/amap-driving.service';
+import { AmapTruckDrivingService, AmapTruckDrivingWrapper } from './services/amap-truck-driving/amap-truck-driving.service';
 import { AmapTransferService, AmapTransferWrapper } from './services/amap-transfer/amap-transfer.service';
 import { AmapWalkingService, AmapWalkingWrapper } from './services/amap-walking/amap-walking.service';
 import { AmapRidingService, AmapRidingWrapper } from './services/amap-riding/amap-riding.service';
@@ -92,7 +93,8 @@ export class NgxAmapModule {
         AmapDrivingService,
         AmapTransferService,
         AmapWalkingService,
-        AmapRidingService
+        AmapRidingService,
+        AmapTruckDrivingService
       ]
     };
   }
@@ -131,5 +133,7 @@ export {
   AmapWalkingService,
   AmapWalkingWrapper,
   AmapRidingService,
-  AmapRidingWrapper
+  AmapRidingWrapper,
+  AmapTruckDrivingService,
+  AmapTruckDrivingWrapper
 };