+++ projects/ngx-amap/src/types/interface/index.ts
@@ -24,6 +24,7 @@ export * from './overlays/ellipse-options.interface';
 export * from './overlays/rectangle-options.interface';
 export * from './layers/heatmap-options.interface';
 export * from './driving/driving-options.interface';
+export * from './driving/truck-driving-options.interface';
 export * from './transfer/transfer-options.interface';
 export * from './walking/walking-options.interface';
 export * from './riding/riding-options.interface';