+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### New
 + 支持路径规划：
   + AMap.Driving 驾车路线规划服务：AmapDrivingService
+  + AMap.TruckDriving 货车路线规划服务：AmapTruckDrivingService
   + AMap.Transfer 公交换乘路线规划服务：AmapTransferService
   + AMap.Walking 步行路线规划服务：AmapWalkingService
   + AMap.Riding 骑行路线规划服务：AmapRidingService